+++ CHANGELOG.md
@@ -3,8 +3,9 @@
 ## 1.0.1 - 2020-12-11
 
 ### Added
-- Support for "Automatic Return" using PKCE
-- `returnUrl()` function for button to redirect the user to a specific URL on return
+- Added support for "Automatic Return" using PKCE
+- Added `returnUrl()` function for button to redirect the user to a specific URL on return
+- Added `getLogoutUrl()` function to get a link that removes both the Vipps-session and the Craft-session.
 
 ## 1.0.0 - 2020-11-14